add more styling, add auth check for editing deleting and creating comments and posts, adjust time display

+++ controllers/posts.js
@@ -34,10 +34,14 @@ function newPost (req, res) {
 
 async function createPost (req, res) {
   try {
-    req.body.user = req.user._id
-    const post = await Post.create(req.body);
-    //redirect to the created post
-    res.redirect('/posts/all');
+    if (!req.user) {
+      res.redirect('posts/all')
+    } else {
+      req.body.user = req.user._id;
+      const post = await Post.create(req.body);
+      //redirect to the created post
+      res.redirect('/posts/all');
+    }
   } catch (err) {
     console.log(err);
     res.render('posts/new', { errorMsg: err.message });
@@ -45,16 +49,20 @@ async function createPost (req, res) {
 }
 
 async function showMy (req, res) {
-  const currentUserId = req.user ? req.user._id : null;
-  const posts = await Post.find({user: currentUserId}).sort({createdAt: -1});
-  const currentUser = req.user ? req.user: null;
+  if (!req.user) {
+    res.redirect('/posts/all');
+  } else {
+    const currentUserId = req.user ? req.user._id : null;
+    const posts = await Post.find({user: currentUserId}).sort({createdAt: -1});
+    const currentUser = req.user ? req.user: null;
 
-  res.render('posts/my', { 
-    title: 'My Posts', 
-    posts, 
-    currentUserId,
-    currentUser 
-  });
+    res.render('posts/my', { 
+      title: 'My Posts', 
+      posts, 
+      currentUserId,
+      currentUser 
+    });
+  }
 }
 
 async function showOne (req, res) {
@@ -84,9 +92,15 @@ async function showOne (req, res) {
 
 async function deleteOnePost(req, res) {
   try {
-    const post = await Post.findById(req.params.id);
-    await Post.findOneAndDelete({_id: post._id});
-    res.redirect('/posts/all');
+    if (!req.user) {
+      res.redirect('/posts/all');
+    } else {
+      const post = await Post.findById(req.params.id);
+      if (post.user.toString() === req.user._id.toString()){
+        await Post.findOneAndDelete({_id: post._id});
+      }
+      res.redirect('/posts/all');
+    }
   } catch(err) {
     console.error(err);
     res.redirect('/posts/all');
@@ -94,21 +108,39 @@ async function deleteOnePost(req, res) {
 }
 
 async function editPost(req, res) {
-  const currentUserId = req.user ? req.user._id : null;
-  const post = await Post.findById(req.params.id);
-  res.render('posts/edit', { title: post.title, post, currentUserId });
+  if (!req.user) {
+    res.redirect('/posts/all');
+  } else {
+    const currentUserId = req.user ? req.user._id : null;
+    const post = await Post.findById(req.params.id);
+    if (currentUserId.toString() === post.user.toString()) {
+      res.render('posts/edit', { title: post.title, post, currentUserId });
+    } else {
+      res.redirect('/posts/all');
+    }
+    
+  }
 }
 
 async function updatePost(req, res) {
   try {
-    const post = await Post.findById(req.params.id);
-    post.title = req.body.title;
-    post.contents = req.body.contents;
-    await post.save();
-    res.redirect('/posts/all');
+    if (!req.user) {
+      res.redirect('/posts/all');
+    } else {
+      const post = await Post.findById(req.params.id);
+      if (req.user._id.toString() === post.user.toString()) {
+        post.title = req.body.title;
+        post.contents = req.body.contents;
+        await post.save();
+        res.redirect(`/posts/${req.params.id}`);
+      } else {
+        res.redirect('/posts/all');
+      }
+    }
   } catch(err) {
     console.log(err);
-    res.redirect('/posts/all');
+    res.redirect(`/posts/${req.params.id}`);
+    
   }
 }
